+++ src/builder.ts
@@ -1,4 +1,5 @@
 import { Database, Table } from './database'
+import { BooleanColumn, Column } from './columns'
 import { Condition } from './models'
 import { Binder, BinderStore } from './binder'
 import { ASTERISK, Distinct, All } from './singletoneConstants'
@@ -25,7 +26,8 @@ export type BuilderData = {
   //TODO: make table "FromItemInfo" array
   table?: Table,
   distinct: ''|' DISTINCT'|' ALL'
-  whereParts: (LogicalOperator|Condition|Parenthesis)[],
+  whereParts: (LogicalOperator|Condition|Parenthesis|BooleanColumn)[],
+  groupByItems: Column[],
   orderByItemInfos: OrderByItemInfo[],
   limit?: null|number|Binder|All,
   offset?: number|Binder,
@@ -43,8 +45,9 @@ export class Builder {
       selectItemInfos: [],
       distinct: '',
       whereParts: [],
+      groupByItems: [],
       orderByItemInfos: [],
-      binderStore: BinderStore.getInstance(),
+      binderStore: new BinderStore(),
       option: fillUndefinedOptionsWithDefault(option ?? {}),
     }
     this.rootStep = new Step(this.data)
@@ -55,7 +58,6 @@ export class Builder {
   public select(...items: (Distinct|All|SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
     if (items[0] instanceof Distinct) {
       if (items.length <= 1) throw new Error('Select step must have at least one parameter after DISTINCT')
-      this.rootStep.cleanUp()
       items.shift() //remove first item the DISTINCT item
       Builder.throwIfMoreThanOneDistinctOrAll(items)
       const newItems = items as unknown[] as (SelectItemInfo|SelectItem|PrimitiveType)[]
@@ -63,35 +65,32 @@ export class Builder {
     }
 
     if (items[0] instanceof All) {
-      this.rootStep.cleanUp()
       items.shift() //remove first item the ALL item
       Builder.throwIfMoreThanOneDistinctOrAll(items)
       const newItems = items as (SelectItemInfo|SelectItem|PrimitiveType)[]
       return this.rootStep.selectAll(...newItems)
     }
 
-    this.rootStep.cleanUp()
     Builder.throwIfMoreThanOneDistinctOrAll(items)
     const newItems = items as (SelectItemInfo|SelectItem|PrimitiveType)[]
     return this.rootStep.select(...newItems)
   }
 
   public selectDistinct(...items: (SelectItem|PrimitiveType)[]): SelectStep {
-    //Note: the cleanup needed as there is only one "select" step in the chain that we start with
-    this.rootStep.cleanUp()
     return this.rootStep.selectDistinct(...items)
   }
 
   public selectAll(...items: (SelectItem|PrimitiveType)[]): SelectStep {
-    //Note: the cleanup needed as there is only one "select" step in the chain that we start with
-    this.rootStep.cleanUp()
     return this.rootStep.selectAll(...items)
   }
 
   public selectAsteriskFrom(table: Table): FromStep {
-    //Note: the cleanup needed as there is only one "select" step in the chain that we start with
-    this.rootStep.cleanUp()
     return this.rootStep.select(ASTERISK).from(table)
+  }
+
+  public cleanUp(): Builder {
+    this.rootStep.cleanUp()
+    return this
   }
 
   private static throwIfMoreThanOneDistinctOrAll(items: (Distinct|All|SelectItemInfo|SelectItem|PrimitiveType)[]) {